requirements.txt
Django~=5.0.7
djangorestframework~=3.15.2
aiohttp~=3.9.5
beautifulsoup4~=4.12.3
selenium~=4.23.1
python-dotenv~=1.0.1